2009-12-10	Ralf Corsépius <ralf.corsepius@rtems.org>
	* console/console.c:
	Reflect changes to rtems_termios_callbacks->write.

+++ c/src/lib/libbsp/sparc/erc32/console/console.c
@@ -281,7 +281,7 @@ void console_outbyte_interrupt(
  *
  */
 
-int console_write_support (int minor, const char *buf, int len)
+ssize_t console_write_support (int minor, const char *buf, size_t len)
 {
   int nwrite = 0;
