+++ c/src/lib/libbsp/sparc/leon3/console/console.c
@@ -59,7 +59,7 @@ int console_inbyte_nonblocking( int port );
  *
  */
 
-int console_write_support (int minor, const char *buf, int len)
+ssize_t console_write_support (int minor, const char *buf, size_t len)
 {
   int nwrite = 0;
